use Mainflux entities
